Displaying dynamic error message when signup validation fails. Also fixed bug that crashes app if a user attempts to logout immediately following registration (app was trying to delete a non-existent session object)

+++ SnailMail/ViewProfileViewController.swift
@@ -70,15 +70,16 @@ class ViewProfileViewController: UIViewController, UITableViewDelegate, UITableV
         let fetchRequest = NSFetchRequest(entityName: "Session")
         var error: NSError?
         
-        let fetchedResults = managedContext.executeFetchRequest(fetchRequest, error: &error) as? [NSManagedObject]
+        let fetchedResults = managedContext.executeFetchRequest(fetchRequest, error: &error)! as [AnyObject]
         
-        let session = fetchedResults![0] as NSManagedObject
-        
-        managedContext.deleteObject(session)
+        if fetchedResults.count > 0 {
+            if let session = fetchedResults[0] as? NSManagedObject {
+                managedContext.deleteObject(session)
+            }
+        }
         
     }
 
-    
     func getOutbox() {
         DataManager.getMyOutbox( { (error, result) -> Void in
             if error != nil {
